【カ】【L】【データINSERT用】【All】仮ロジック＆仮テストの作成 #115
- マッピングミスの修正
- dto -> usecaseの依存を削除
