+++ code/cardApiCall/infoInsert/usecase/tcgapi/tcg.go
@@ -5,38 +5,21 @@ import (
 	"encoding/json"
 	"errors"
 
+	"atomisu.com/ocg-statics/infoInsert/dto/cardrecord"
 	"atomisu.com/ocg-statics/infoInsert/http"
 	"atomisu.com/ocg-statics/infoInsert/usecase"
 	"github.com/samber/do"
 )
 
-type TcgApiCard struct {
-	Desc                  string   `json:"desc"`
-	Name                  string   `json:"name"`
-	ID                    int64    `json:"id"`
-	Def                   int32    `json:"def"`
-	Atk                   int32    `json:"atk"`
-	Type                  string   `json:"type"`
-	Level                 int32    `json:"level"`
-	Race                  string   `json:"race"`
-	LinkMarkers           []string `json:"linkMarkers"`
-	Attribute             string   `json:"attribute"`
-	LinkVal               int32    `json:"linkVal"`
-	TypeLines             []string `json:"typeline"`
-	HumanReadableCardType string   `json:"humanReadableCardType"`
-	Scale                 int32    `json:"scale"`
-	PendulumText          string   `json:"pendulumText"`
-}
-
 // TcgApiResponse is a response from the TcgRest.
 type TcgApiResponse struct {
-	Data []TcgApiCard `json:"data"`
+	Data []cardrecord.TcgApiCard `json:"data"`
 }
 
 // TcgUseCase is an interface for the TcgUseCase.
 type TcgUseCase interface {
 	usecase.UseCase
-	GetCardInfoByEnName(ctx context.Context, name string) (TcgApiCard, error)
+	GetCardInfoByEnName(ctx context.Context, name string) (cardrecord.TcgApiCard, error)
 }
 
 type tcgUseCaseImpl struct {
@@ -50,22 +33,23 @@ func NewTcgUseCase(i *do.Injector) (TcgUseCase, error) {
 	})
 }
 
-func (t *tcgUseCaseImpl) GetCardInfoByEnName(ctx context.Context, name string) (TcgApiCard, error) {
+func (t *tcgUseCaseImpl) GetCardInfoByEnName(ctx context.Context, name string) (cardrecord.TcgApiCard, error) {
 	tcgRest := http.NewTCGRest()
 	results, err := tcgRest.GetEnInfoByEnName(ctx, name)
 	if err != nil {
-		return TcgApiCard{}, err
+		return cardrecord.TcgApiCard{}, err
 	}
 
 	var response TcgApiResponse
 	err = json.Unmarshal(results, &response)
 	if err != nil {
-		return TcgApiCard{}, err
+		return cardrecord.TcgApiCard{}, err
 	}
 
 	if len(response.Data) == 0 {
-		return TcgApiCard{}, errors.New("no data found")
+		return cardrecord.TcgApiCard{}, errors.New("no data found")
 	}
 
-	return response.Data[0], nil
+	tcgCard := response.Data[0]
+	return tcgCard, nil
 }
